Backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
chromadb
langchain
langchain-google-genai
google-generativeai
httpx
pydantic[email] 